Netflix Userbase Analysis
In my quest to understand how Netflix makes money and who its users are, I explored a dataset from Kaggle. Using tools like Microsoft Excel and Power BI, I uncovered some important insights about Netflix's earnings and user behavior.

Netflix has 2,500 registered users, with almost an equal number of men and women. This shows that Netflix's content is appealing to both genders, ensuring a diverse audience. The most active viewers are between the ages of 26-32. This group likely has the disposable income and lifestyle that fits regular streaming.

Laptops are the most popular choice for streaming Netflix content. This is probably because laptops are versatile and convenient, allowing users to watch their favorite shows and movies from almost anywhere. On the other hand, Smart TVs, though popular in homes, are the least used for streaming Netflix. This might mean that while Smart TVs are great for family or group watching, they may not be the go-to device for individuals.

Netflix's monthly revenue is a substantial $31,271. This income comes from its three subscription plans: Basic, Standard, and Premium. The Basic plan is the most profitable, contributing $12,469 to the total monthly revenue. The Standard plan follows closely, generating $9,537, while the Premium plan brings in $9,229. These figures show that Netflix has a pricing strategy that caters to different user needs and preferences.

The insights from this analysis offer a glimpse into Netflix's user base and revenue. By knowing which plans are the most profitable, which devices are favored, and the demographics of its most active users, Netflix can better tailor its services and marketing strategies to meet the needs of its audience.